Module libra_config::config

Structs

AdmissionControlConfig

BaseConfig

ConsensusConfig

DebugInterfaceConfig

ExecutionConfig

LoggerConfig

MempoolConfig

MetricsConfig

NetworkConfig

NodeConfig
Config pulls in configuration information from the config file. This is used to set up the nodes and configure various parameters. The config file is broken up into sections for each module so that only that module can be passed around

NodeConfigHelpers

StateSyncConfig

StorageConfig

VMConfig
Holds the VM configuration, currently this is only the publishing options for scripts and modules, but in the future this may need to be expanded to hold more information.

Enums

ConsensusProposerType

RoleType

VMPublishingOption
Defines and holds the publishing policies for the VM. There are three possible configurations:

Traits

PersistableConfig