Module libra_config::keys

This file implements a KeyPair data structure.

The point of a KeyPair is to deserialize a private key into a structure that will only allow the private key to be moved out once (hence providing good key hygiene) while allowing access to the public key part forever.

The public key part is dynamically derived during deserialization, while ignored during serialization.

Structs

KeyPair
A KeyPair has a private key that can only be taken once, and a public key that can be cloned ad infinitum.