Module libra_config::config

Structs

AdmissionControlConfig

BaseConfig

ConsensusConfig

DebugInterfaceConfig

ExecutionConfig

GitHubConfig

KeyManagerConfig

LeaderReputationConfig

LoggerConfig

MempoolConfig

MetricsConfig

NetworkConfig

NetworkKeyPairs

NetworkPeerInfo

NetworkPeersConfig

NodeConfig
Config pulls in configuration information from the config file. This is used to set up the nodes and configure various parameters. The config file is broken up into sections for each module so that only that module can be passed around

OnDiskStorageConfig

ParseRoleError

PeerNetworkId
Identifier of a node, represented as (network_id, peer_id)

RemoteExecutionService

RemoteService

RootPath

RpcConfig

SafetyRulesConfig

SecureConfig

SeedPeersConfig

StateSyncConfig

StorageConfig

TestConfig

UpstreamConfig

VaultConfig

Enums

ConsensusProposerType

ConsensusType

DiscoveryMethod

ExecutionCorrectnessService
Defines how execution correctness should be run

RoleType

SafetyRulesService
Defines how safety rules should be executed

SecureBackend

Constants

DEFAULT_JSON_RPC_PORT

Traits

PersistableConfig

Type Definitions

NetworkId
In general, a network ID is a PeerId that this node uses to uniquely identify a network it belongs to. This is equivalent to the peer_id field in the NetworkConfig of this NodeConfig